changed App architecture created packages
