+++ app/src/main/java/sarveshchavan777/Views/GameWon.java
@@ -1,9 +1,11 @@
-package sarveshchavan777.triviaquiz;
+package sarveshchavan777.Views;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+
+import sarveshchavan777.triviaquiz.R;
 
 public class GameWon extends Activity {
