+++ app/src/main/java/sarveshchavan777/Views/Time_Up.java
@@ -1,45 +1,49 @@
-package sarveshchavan777.triviaquiz;
+package sarveshchavan777.Views;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Typeface;
+import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TextView;
 
+import info.hoang8f.widget.FButton;
+import sarveshchavan777.triviaquiz.R;
 
-public class PlayAgain extends Activity {
-
-    Button playAgain;
-    TextView wrongAnsText;
+public class Time_Up extends AppCompatActivity {
+    FButton playAgainButton;
+    TextView timeUpText;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.play_again);
-       //Initialize
-        playAgain = (Button) findViewById(R.id.playAgainButton);
-        wrongAnsText = (TextView)findViewById(R.id.wrongAns);
+        setContentView(R.layout.activity_time__up);
+        //Initialize
+        playAgainButton = (FButton)findViewById(R.id.playAgainButton);
+        timeUpText = (TextView)findViewById(R.id.timeUpText);
 
         //play again button onclick listener
-        playAgain.setOnClickListener(new View.OnClickListener() {
+        playAgainButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(PlayAgain.this, MainGameActivity.class);
+                Intent intent = new Intent(Time_Up.this, MainGameActivity.class);
                 startActivity(intent);
                 finish();
+
+
             }
         });
 
+
         //Setting typefaces for textview and button - this will give stylish fonts on textview and button
         Typeface typeface = Typeface.createFromAsset(getAssets(),"fonts/shablagooital.ttf");
-        playAgain.setTypeface(typeface);
-        wrongAnsText.setTypeface(typeface);
+        timeUpText.setTypeface(typeface);
+        playAgainButton.setTypeface(typeface);
     }
 
     @Override
     public void onBackPressed() {
+        super.onBackPressed();
         finish();
     }
 }
