+++ app/src/main/res/layout-normal/activity_time__up.xml
@@ -7,7 +7,7 @@
     android:background="@color/backgroundColor"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context="sarveshchavan777.triviaquiz.Time_Up">
+    tools:context="sarveshchavan777.triviaquiz.sarveshchavan777.Views.Time_Up">
 
     <TextView
         android:id="@+id/timeUpText"
