+++ app/src/main/res/layout-small/activity_home_screen.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:gravity="center"
     android:background="@color/backgroundColor"
-    tools:context="sarveshchavan777.triviaquiz.HomeScreen">
+    tools:context="sarveshchavan777.triviaquiz.sarveshchavan777.Views.HomeScreen">
 <TextView
     android:id="@+id/tQ"
     android:layout_width="match_parent"
